Refactor : 메인 냉털 레시피 컴포넌트 비즈니스 로직 분리
