+++ src/ReactQuery/queryKey.ts
@@ -5,6 +5,7 @@ export const queryKey = {
   recipes: "recipes",
   bestRecipes: "bestRecipes",
   newestRecipes: "newestRecipes",
+  singleRecipes: "singleRecipes",
   recipe: "recipe",
   comments: "comments",
   comment: "comment",
