+++ src/hooks/useRecipesQuery.ts
@@ -1,7 +1,11 @@
 import { useQuery } from "@tanstack/react-query";
 import { Recipe } from "../types";
 import { queryKey } from "../ReactQuery/queryKey";
-import { getRecipesByLastest, getRecipesByPopularity } from "../app/api/recipe";
+import {
+  getRecipesByLastest,
+  getRecipesByPopularity,
+  getRecipesBySingle,
+} from "../app/api/recipe";
 
 export const useBestRecipesQuery = () => {
   const fallback = [] as Recipe[];
@@ -23,6 +27,18 @@ export const useNewestRecipesQuery = () => {
     isLoading,
     isError,
   } = useQuery<Recipe[]>([queryKey.newestRecipes], getRecipesByLastest);
+
+  return { data, isLoading, isError };
+};
+
+export const useSingleRecipesQuery = () => {
+  const fallback = [] as Recipe[];
+
+  const {
+    data = fallback,
+    isLoading,
+    isError,
+  } = useQuery<Recipe[]>([queryKey.singleRecipes], getRecipesBySingle);
 
   return { data, isLoading, isError };
 };
